% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manipulation.R
\name{AddAllTeamsDb}
\alias{AddAllTeamsDb}
\title{Add team metadata to the database}
\usage{
AddAllTeamsDb()
}
\description{
Add all teams to the database. The teams will get stored in the database in the 'teams' table.
Will not write if teams already exist in the database!
}
\examples{
SetDbPath(example = TRUE)
AddAllTeamsDb()

}
