% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{EventsExists}
\alias{EventsExists}
\title{Check if the events table exists, returns boolean}
\usage{
EventsExists()
}
\description{
Check if the events table exists, returns boolean
}
\keyword{internal}
