% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{GetDbPath}
\alias{GetDbPath}
\title{Returns the current path to the database file.}
\usage{
GetDbPath()
}
\value{
String containing the path to the database.
}
\description{
Returns the current path to the database file.
}
\examples{
GetDbPath()

}
