% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{GetPlayerId}
\alias{GetPlayerId}
\title{Gets a player id from their name. Will only work for players that were active in a game that
has already been added to the database.}
\usage{
GetPlayerId(player_name)
}
\arguments{
\item{player_name}{String, players full name.}
}
\value{
Int, player id number.
}
\description{
Gets a player id from their name. Will only work for players that were active in a game that
has already been added to the database.
}
\examples{
SetDbPath()
AddGameEvents(2019020001)
GetPlayerId("John Tavares")

}
