% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{GetPlayerStats}
\alias{GetPlayerStats}
\title{WIP - Get advanced statistics for player_id on team_id in a list of games.
The current stats returned are:
- Corsi For
- Corsi Against
- Shots}
\usage{
GetPlayerStats(player_id, game_ids, team_id)
}
\arguments{
\item{player_id}{Player ID number.}

\item{game_ids}{List of game ids to check. Must already be in the database.}

\item{team_id}{The ID of the team the player plays for.}
}
\value{
Dataframe containing a row of stats for even strength and for all situations.
}
\description{
WIP - Get advanced statistics for player_id on team_id in a list of games.
The current stats returned are:
- Corsi For
- Corsi Against
- Shots
}
\examples{
SetDbPath()
AddGameEvents(2019020001)
GetPlayerStats(8475166, 2019020001, 10)

}
