% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{GetTeamId}
\alias{GetTeamId}
\title{Retrieve the team ID using the abbreviation, or full name of the team.}
\usage{
GetTeamId(team_name)
}
\arguments{
\item{team_name}{String containing either abbreviation or full name.}
}
\value{
Int, team ID number.
}
\description{
Retrieve the team ID using the abbreviation, or full name of the team.
}
\examples{

AddAllTeamsDb()
GetTeamId("TOR")
GetTeamId("Toronto Maple Leafs")

}
