% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_scrape.R
\name{QueryDb}
\alias{QueryDb}
\title{Send a SQL query to the database.}
\usage{
QueryDb(query)
}
\arguments{
\item{query}{A string containing the SQL query.}
}
\description{
Send a SQL query to the database.
}
\examples{
QueryDb("SELECT * FROM events")
QueryDb("SELECT player_id, player_fullName FROM events WHERE game_id=2019020001")
}
