% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{SetDbPath}
\alias{SetDbPath}
\title{Sets the current path to the database file.}
\usage{
SetDbPath(db_path)
}
\arguments{
\item{db_path}{A string containing the path to the db file.}
}
\value{
String containing the path to the database.
}
\description{
Sets the current path to the database file.
}
\examples{
SetDbPath(system.file("extdata", "nhl.sqlite", package = "nhlscrape"))

}
