% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_scrape.R
\name{AddGameEvents}
\alias{AddGameEvents}
\title{Add all events from a 'game_id' to the 'events' table}
\usage{
AddGameEvents(game_ids)
}
\arguments{
\item{game_ids}{List of identifying numbers for the game. Use GetGameIds to find.}
}
\description{
Add all events from a 'game_id' to the 'events' table
}
\examples{
AddGameEvents(2019020001)

}
