% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manipulation.R
\name{AddTeamRoster}
\alias{AddTeamRoster}
\title{Add a teams roster to the 'roster' table for season. Can be accessed with a primary key
of 'season_teamid_personid'.}
\usage{
AddTeamRoster(team_id, season)
}
\arguments{
\item{team_id}{Identity number for the team. Use GetTeamId to find.}

\item{season}{A year range you want to add.}
}
\description{
Add a teams roster to the 'roster' table for season. Can be accessed with a primary key
of 'season_teamid_personid'.
}
\examples{
SetDbPath()
AddTeamRoster(10, 20192020)

}
