% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_manipulation.R
\name{AddTeamRoster}
\alias{AddTeamRoster}
\title{Adds rosters for team id and a specific season}
\usage{
AddTeamRoster(team_id, season)
}
\arguments{
\item{team_id}{Int, identity number for the team. Use GetTeamId to find}

\item{season}{Int, A year range you want to add}
}
\description{
Add a teams roster to the 'roster' table for season. Can be accessed with a primary key
of 'season_teamid_personid'
}
\examples{
SetDbPath(example = TRUE)
AddTeamRoster(10, 20192020)

}
