% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_scraping.R
\name{GetApiJson}
\alias{GetApiJson}
\title{Call the nhl api and parse the retun}
\usage{
GetApiJson(call)
}
\arguments{
\item{call}{character, contains the api query}
}
\value{
list, contains all the data in the api call
}
\description{
Send a call to the 'NHL' api and get the result back in a list
}
\examples{
\donttest{
GetApiJson("game/2019020001/feed/live")
}

}
