% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{GetDbPath}
\alias{GetDbPath}
\title{Retrieve database location on system}
\usage{
GetDbPath()
}
\value{
Character, system path to the database.
}
\description{
Returns the current path to the local database file
}
\examples{
GetDbPath()

}
