% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{GetPlayerStats}
\alias{GetPlayerStats}
\title{WIP - Get advanced statistics for player_id on team_id in a list of games.}
\usage{
GetPlayerStats(player_id, game_ids, team_id)
}
\arguments{
\item{player_id}{Int, player ID number}

\item{game_ids}{List, game ids to check. Must already be in the database}

\item{team_id}{Int, the ID of the team the player plays for}
}
\value{
List, contains a row of stats for even strength and for all situations
}
\description{
This function will search all games in game_ids and return a list of stats for
the player id selected. The current stats returned are:
- Shots
- Goals
- Corsi
- Fenwick
}
\examples{
SetDbPath(example = TRUE)
AddGameEvents(2019020001)
GetPlayerStats(8475166, 2019020001, 10)

}
