% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{QueryDb}
\alias{QueryDb}
\title{Send a SQL query to the database. Retruns the SQL result as a data.frame. Useful for
seeing the structure of the database for construction of queries.}
\usage{
QueryDb(query)
}
\arguments{
\item{query}{A string containing the SQL query.}
}
\value{
Data.frame containing the SQL query result.
}
\description{
Send a SQL query to the database. Retruns the SQL result as a data.frame. Useful for
seeing the structure of the database for construction of queries.
}
\examples{
SetDbPath()
AddGameEvents(2019020001)
QueryDb("SELECT * FROM events")
QueryDb("SELECT result_description FROM events WHERE game_id=2019020001 AND player_id=8475166")

}
