% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{QueryDb}
\alias{QueryDb}
\title{Send query to database}
\usage{
QueryDb(query)
}
\arguments{
\item{query}{A string containing the SQL query}
}
\value{
List, contains the SQL query result.
}
\description{
Send a SQL query to the database. Retruns the SQL result as a data.frame. Useful for
seeing the structure of the database for construction of queries
}
\examples{
SetDbPath(example = TRUE)
AddGameEvents(2019020001)
QueryDb("SELECT * FROM events")
QueryDb("SELECT result_description FROM events WHERE game_id=2019020001 AND player_id=8475166")

}
