% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{SetDbPath}
\alias{SetDbPath}
\title{Set database location on system}
\usage{
SetDbPath(
  db_path = system.file("extdata", "nhl.sqlite", package = "nhlscrape")
)
}
\arguments{
\item{db_path}{Character, contains the system path to the db file}
}
\value{
Character, contains the path to the database.
}
\description{
Sets the path to the database file. This function must be run to allow operation
of other functions which require saving to a database
}
\examples{
\dontrun{
SetDbPath("C:/Users/Adam/Documents/nhl.sqlite")
}

}
