% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{SetDbPath}
\alias{SetDbPath}
\title{Sets the current path to the database file. This can be done to allow
remote storage or different database files.
Default is nhlscrape/extdata/nhl.sqlite.}
\usage{
SetDbPath(
  db_path = system.file("extdata", "nhl.sqlite", package = "nhlscrape")
)
}
\arguments{
\item{db_path}{A string containing the path to the db file.}
}
\value{
String containing the path to the database.
}
\description{
Sets the current path to the database file. This can be done to allow
remote storage or different database files.
Default is nhlscrape/extdata/nhl.sqlite.
}
\examples{
\dontrun{
SetDbPath("C:/Users/Adam/Documents/nhl.sqlite")
}

}
