% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_retrieval.R
\name{SetDbPath}
\alias{SetDbPath}
\title{Set database location on system}
\usage{
SetDbPath(db_path = "nhl.sqlite", example = FALSE)
}
\arguments{
\item{db_path}{Character, contains the system path to the db file}

\item{example}{boolean, set to FALSE by default, if TRUE it will
copy the example db file and set the path pointing to it}
}
\value{
Character, contains the path to the database.
}
\description{
Sets the path to the database file. This function must be run to allow operation
of other functions which require saving to a database
}
\examples{
SetDbPath(example = TRUE)

}
